backend/requirements.txt
django-storages[azure]==1.9.1
label-studio==1.5
django==3.2.13
djangorestframework==3.13.1
django-dotenv==1.4.2
drf-yasg==1.20.0
djoser==2.1.0
django-cors-headers==3.6.0
django-import-export==2.8.0
psycopg2-binary==2.9.1
lxml==4.9.1
mosestokenizer==1.2.1
indic-nlp-library==0.81
django-smtp-ssl==1.0
gunicorn==20.1.0